# docs/openapi/admin-v1.yaml
openapi: 3.1.0
info:
  title: Assist Platform Admin API
  version: 1.0.0-increment-9
  description: |
    Versioned Admin API for Assist RIC and trusted management clients (CORE-011).
    Increment 9 completes Phase 1 foundation with RIC contract tests.
    Increment 8 adds audit read, search-gap analytics and MFA verify scaffold.
    Increment 7 adds draft/import submission, validation and staging for RIC.
    Increment 6 adds recycle bin list, restore and permanent purge.
    Increment 5 adds audited provider/stay writes and lifecycle transitions.
    Increment 4 adds read-only providers and stays with cursor pagination.
    Increment 3 adds service accounts, machine token exchange and scope enforcement.
    Increment 2 ships human login, refresh rotation, logout, me and sessions.
    Not a supported production partner API until MFA enforcement and Quality Gate evidence.
servers:
  - url: https://{host}/api/v1/admin
    variables:
      host:
        default: vanassist.com.au
...
paths:
...
  /providers:
    get:
      tags: [Providers]
      operationId: adminProvidersIndex
      summary: List providers in the current brand scope
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/QueryQ'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/Town'
        - $ref: '#/components/parameters/State'
      responses:
        '200':
          description: Provider collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderCollection'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing providers:read scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [Providers]
      operationId: adminProvidersStore
      summary: Create a provider in the current brand scope
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderCreateRequest'
      responses:
        '201':
          description: Provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderDetailResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing providers:write scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /providers/{id}:
    get:
      tags: [Providers]
      operationId: adminProvidersShow
      summary: Provider detail in the current brand scope
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Provider detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderDetailResponse'
        '404':
          description: Not found in brand scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Providers]
      operationId: adminProvidersUpdate
      summary: Update provider and brand listing fields
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderPatchRequest'
      responses:
        '200':
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderDetailResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing providers:write scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found in brand scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Providers]
      operationId: adminProvidersDestroy
      summary: Soft-delete a provider (reason required)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleDeleteRequest'
      responses:
        '200':
          description: Provider soft-deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '401':
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Missing lifecycle:write scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
    QueryQ:
      name: q
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      description: DB status or lifecycle alias (published, unpublished, pending_review, inactive)
      schema:
        type: string
    Town:
      name: town
      in: query
      schema:
        type: string
    State:
      name: state
      in: query
      description: State id, abbreviation, or name
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque access token from /auth/login, /auth/refresh or /auth/token. Only hashes are stored.
  schemas:
...
    ProviderCollection:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProviderSummary'
        meta:
          type: object
          additionalProperties: true
        links:
          type: object
          additionalProperties: true
    ProviderDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/ProviderDetail'
    ProviderSummary:
      type: object
      required: [id, business_name, lifecycle]
      properties:
        id:
          type: string
        business_name:
          type: string
        slug:
          type: string
        status:
          type: string
        lifecycle:
          type: string
        town:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
    ProviderDetail:
      allOf:
        - $ref: '#/components/schemas/ProviderSummary'
        - type: object
          additionalProperties: true
...
    ProviderCreateRequest:
      type: object
      required: [business_name]
      properties:
        business_name:
          type: string
        display_name:
          type: string
        status:
          type: string
          enum: [draft, pending]
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        public_email:
          type: string
        public_phone:
          type: string
        website:
          type: string
        description:
          type: string
        service_model:
          type: string
          enum: [mobile, workshop, both]
        street_address:
          type: string
        base_town_id:
          type: integer
        region_id:
          type: integer
    ProviderPatchRequest:
      type: object
      properties:
        business_name:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
        public_email:
          type: string
        public_phone:
          type: string
        website:
          type: string
        description:
          type: string
        service_model:
          type: string
          enum: [mobile, workshop, both]
        street_address:
          type: string
        base_town_id:
          type: integer
        region_id:
          type: integer
        display_name:
          type: string
        search_visible:
          type: boolean
        listing_status:
          type: string
          enum: [draft, pending, active, suspended, rejected]
...
    LifecycleDeleteRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 3
    DeleteResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, deleted, reason]
          properties:
            id:
              type: string
            deleted:
              type: boolean
            reason:
              type: string
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
        meta:
          type: object
          additionalProperties: true